Fix delegation bonus ABI mismatch and harden admin scripts
- Fix ILazyDelegateRegistry.getSerialsDelegatedTo return type from int64[]
  to uint256[] in ILazyLottoPoolManager.sol, matching the actual
  LazyDelegateRegistry implementation — this caused calculateBoost to
  silently fail on ABI decode, never applying NFT delegation bonuses.

- Rewrite removePrizes.js: pass both required params (poolId, prizeIndex),
  loop with swap-and-pop removal, fix prize display to use actual struct
  fields, scale gas by max NFT count, auto-approve HBAR allowance for
  royalty dust transfers, and retry on transient UNKNOWN (21) errors.

- Rewrite buyAndRedeemEntry.js to mint in batches of 10 per transaction
  instead of one oversized tx that exceeds gas limits for large counts.

+++ contracts/interfaces/ILazyLottoPoolManager.sol
@@ -236,5 +236,5 @@ interface ILazyDelegateRegistry {
     function getSerialsDelegatedTo(
         address _user,
         address _token
-    ) external view returns (int64[] memory serials);
+    ) external view returns (uint256[] memory serials);
 }
